backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
jinja2==3.1.4
python-multipart==0.0.9
alembic==1.13.2
pytest==8.3.3
httpx==0.27.0
python-jose[cryptography]==3.3.0
